requirements.txt
altgraph==0.17.4
anyio==4.11.0
certifi==2025.11.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==25.0
pefile==2023.2.7
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
python-dotenv==1.2.1
python-telegram-bot==22.5
pywin32-ctypes==0.2.3
setuptools==80.9.0
sniffio==1.3.1
